requirements.txt


# Vector database
chromadb>=0.4.0

# Document processing
PyMuPDF>=1.23.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
Pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
python-pptx>=0.6.21
striprtf>=0.0.26
markdown>=3.4.0
PyYAML>=6.0

# Environment and configuration
configparser>=5.3.0

# API requirements
Flask==2.3.3
Flask-CORS==4.0.0
uvicorn>=0.23.0
fastapi>=0.104.0
pydantic>=2.0.0
sse-starlette>=1.0.0

# Database
elasticsearch==8.9.0
mysql-connector-python>=8.0.0

# LLM and AI
ollama>=0.1.0
langchain>=0.0.267
langchain-text-splitters>=0.0.1


# Utilities
uuid>=1.30
typing>=3.7.4
python-dotenv>=1.0.0
sentence-transformers